requirements.txt
# Core libraries
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.0
tensorboard>=2.5.0
tensorboardX>=2.1

# Data processing and analysis
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pillow>=8.0.0
joblib>=1.0.0
tqdm>=4.60.0
